+++ IO/vtkXMLPImageDataReader.h
@@ -47,8 +47,8 @@ protected:
   vtkXMLPImageDataReader();
   ~vtkXMLPImageDataReader();
   
-  float Origin[3];
-  float Spacing[3];
+  double Origin[3];
+  double Spacing[3];
   
   vtkImageData* GetPieceInput(int index);
